Make View/Edit point to proper product scope (Refs #569).
When creating a ticket through the QCT dialog for a product when in the scope of another product, the //View/Edit// link would not be properly product-scoped. For example, when creating a ticket associated with product B when in the scope of product A, the `href` for the //View/Edit// link would be `/products/productA/ticket/id`, but should be `/products/productB/ticket/id`.

Changes have been tested on the desktop and mobile browsers.

git-svn-id: https://svn.apache.org/repos/asf/bloodhound/trunk@1500163 13f79535-47bb-0310-9956-ffa450edef68