0.8dev: Don't try to add duplicate ticket relation when the `Ticket.duplicate` attribute is not present. Refs #710.
The `ITicketManipulator` interface isn't called when batch modifying tickets, so the `duplicate` attribute doesn't get added to the ticket object and ticket validation is not performed. The fix here is a temporary workaround to avoid a traceback, until batch modifications can add a duplicate ticket relation. Patch by Olemis Lang.


git-svn-id: https://svn.apache.org/repos/asf/bloodhound/trunk@1538891 13f79535-47bb-0310-9956-ffa450edef68