0.8dev: Restore border around the //Age// column.
The `browser.css` stylesheet from Trac removes the border from the `td` in the age column. This doesn't affect the border displayed in Trac because there is a border in `trac.css` defined for the `tr`. In `bootstrap.css` though, the border is defined on the `td` element, therefore we need to define the same border styling for the `tr` element.

The alternative would be to modify `browser.css`, which we don't do for now so as to continue to use an unmodified version from the Trac core. If we move to using LESS, this can probably be simplified.


git-svn-id: https://svn.apache.org/repos/asf/bloodhound/trunk@1545220 13f79535-47bb-0310-9956-ffa450edef68