0.8dev: //resolve as duplicate// now works correctly for any action with the `set_resolution` operation. Refs #742.
Previously, the //Duplicate ID// input would not be displayed for any action other than //resolve//. Also the `duplicate` attribute would only be bound in `ITicketManipulator` when the action was named `resolve`.

Now a //Duplicate ID// input will be attached to the workflow action and the `duplicate` attribute will be bound by `ITicketManipulator` for any action with the `set_resolution` operation. This has been tested and functions property when there are multiple actions that specify the `set_resolution` operation.



git-svn-id: https://svn.apache.org/repos/asf/bloodhound/trunk@1560601 13f79535-47bb-0310-9956-ffa450edef68