0.8dev: Fixed error in unit test revealed by [1560601]. Refs #742.
The parent ticket should not have status //closed// since the //resolve// action is being invoked on it. This error was revealed after [1560601] since we are now looking up operations for the //action//, which depends on ticket status. With the default workflow, //leave// and //del_resolution// will be the only operations for a closed ticket.


git-svn-id: https://svn.apache.org/repos/asf/bloodhound/trunk@1560613 13f79535-47bb-0310-9956-ffa450edef68